+++ CHANGELOG.md
@@ -11,6 +11,8 @@ FlatLaf Change Log
 - Fixed selection background of checkbox in table cell.
 - Fixed jittery submenu rendering on Mac. (issue #10)
 - Hex color values in `.properties` files now must start with a `#` character.
+- SwingX: Fixed too wide border when using date picker as table cell editor.
+  (issue #24)
 - JIDE Common Layer: Fixed `JidePopup` border.